fix bundling
