+++ alchemy/src/cloudflare/bundle/external.ts
@@ -5,13 +5,14 @@ const nodejs_compat = [
   "node:buffer",
   "node:console",
   "node:crypto",
+  "node:debug",
   "node:diagnostics_channel",
   "node:dns",
   "node:events",
   "node:inspector",
   "node:net",
   "node:path",
-  "node:perf_hooks",
+  "node:perf_hooks", // partially supported
   "node:process",
   "node:querystring",
   "node:stream",
